fix(billing): compute tax_amount from the persisted numeric(5,2) rate
Cast the tax_amount calc operand to numeric(5,2) so the stored tax_rate and the
amount it produced can never disagree (defensive — config already rejects >2dp
rates). Strengthen the Customers & Bills read test to assert the repeatable-read,
read-only snapshot isolation alongside the shared-tx-handle check.

+++ db/repositories/billing/customer-bill-tax-item.repository.ts
@@ -49,7 +49,12 @@ export const customerBillTaxItemRepository = {
         cb.period_partition,
         ${input.category},
         ${input.rate}::numeric(5, 2),
-        round(cb.subtotal * ${input.rate}::numeric / 100, 2)
+        -- Compute the amount from the SAME numeric(5,2) rate that is persisted
+        -- in tax_rate above, so the stored rate and the amount it produced can
+        -- never disagree (the config already rejects >2dp rates, so this is
+        -- defensive — a future caller passing a higher-precision rate stays
+        -- consistent). round(...,2) keeps the half-up two-decimal amount.
+        round(cb.subtotal * ${input.rate}::numeric(5, 2) / 100, 2)
       FROM billing.customer_bill AS cb
       WHERE cb.customer_bill_id = ${input.customerBillId}
         AND cb.period_partition = ${input.periodPartition}
